# backend/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Inventory & Sales API",
    "version": "1.0.0",
    "description": "API documentation for the Inventory & Sales system",
    "contact": {}
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "responses": {
      "UnauthorizedError": {
        "description": "Access token is missing or invalid"
      },
      "ForbiddenError": {
        "description": "Access token is invalid"
      }
    }
  },
...
  "servers": [
    {
      "url": "http://localhost:5000"
    }
  ],
  "paths": {
...
    "/v1/api/colors": {
      "get": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "add color",
        "tags": ["color"],
        "description": "add color",
        "operationId": "getColors",
        "responses": {
          "200": {
            "description": "Successfully get color",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "color_code": {
                        "type": "string",
                        "example": "A553"
                      },
                      "fabric_type": {
                        "type": "string",
                        "example": "Katun"
                      },
                      "color": {
                        "type": "string",
                        "example": "Red"
                      },
                      "by_who": {
                        "type": "string",
                        "example": "586949c5-db09-4cc8-8d90-187c0881bf3c"
                      },
                      "createdAt": {
                        "type": "string",
                        "example": "2025-03-19T04:43:11.596Z"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          }
        }
      },
      "post": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "add color",
        "tags": ["color"],
        "description": "add color",
        "operationId": "colorAdd",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "color_code": {
                    "type": "string",
                    "example": "A555"
                  },
                  "fabric_type": {
                    "type": "string",
                    "example": "Poliester"
                  },
                  "color": {
                    "type": "string",
                    "example": "Blue"
                  }
                }
              },
              "example": {
                "color_code": "A555",
                "fabric_type": "Poliester",
                "color": "Blue"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully added color",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Color created"
                    },
                    "colorAdd": {
                      "type": "object",
                      "properties": {
                        "createdAt": {
                          "type": "string",
                          "example": "2025-03-21T06:13:48.106Z"
                        },
                        "color_code": {
                          "type": "string",
                          "example": "A555"
                        },
                        "fabric_type": {
                          "type": "string",
                          "example": "Poliester"
                        },
                        "color": {
                          "type": "string",
                          "example": "Blue"
                        },
                        "by_who": {
                          "type": "string",
                          "example": "586949c5-db09-4cc8-8d90-187c0881bf3c"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Missing or invalid fields",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "color, color_code, and fabric_type are required"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          }
        }
      }
    },
...
    "/v1/api/products/{product_id}": {
      "delete": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "delete product",
        "parameters": [
          {
            "in": "path",
            "name": "product_id",
            "required": true,
            "schema": {
              "type": "string",
              "minimum": 1,
              "example": "9743e0d7-a8a6-4c73-a352-3872254ec947"
            },
            "description": "The product ID"
          }
        ],
        "tags": ["product"],
        "description": "delete product",
        "operationId": "deleteProduct",
        "responses": {
          "200": {
            "description": "Successfully delete product",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Product deleted"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          }
        }
      }
    },
    "/v1/api/products/stock": {
      "post": {
        "summary": "add stock",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": ["stock"],
        "description": "add stock",
        "operationId": "addStock",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "product_id": {
                    "type": "string",
                    "example": "049e2d7c-7eb5-4e3b-9bf6-ce93333dde4f"
                  },
                  "price": {
                    "type": "number",
                    "example": 10200
                  },
                  "quantity": {
                    "type": "number",
                    "example": 10
                  }
                }
              },
              "example": {
                "product_id": "049e2d7c-7eb5-4e3b-9bf6-ce93333dde4f",
                "price": 10200,
                "quantity": 10
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully add stock",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Stock added"
                    },
                    "batch": {
                      "type": "object",
                      "properties": {
                        "status": {
                          "type": "string",
                          "example": "new"
                        },
                        "id": {
                          "type": "string",
                          "example": "KAT25_031742559499"
                        },
                        "product_id": {
                          "type": "string",
                          "example": "049e2d7c-7eb5-4e3b-9bf6-ce93333dde4f"
                        },
                        "price": {
                          "type": "integer",
                          "example": 10200
                        },
                        "quantity": {
                          "type": "integer",
                          "example": 10
                        },
                        "by_who": {
                          "type": "string",
                          "example": "586949c5-db09-4cc8-8d90-187c0881bf3c"
                        },
                        "createdAt": {
                          "type": "string",
                          "example": "2025-03-19T04:43:11.596Z"
                        },
                        "updatedAt": {
                          "type": "string",
                          "example": "2025-03-19T04:43:11.596Z"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          }
        }
      }
    },
    "/v1/api/products/reduce-stock": {
      "post": {
        "summary": "reduce stock",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": ["stock"],
        "description": "reduce stock",
        "operationId": "reduceStock",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "product_id": {
                    "type": "string",
                    "example": "f1498e34-b5fc-42bf-8fa0-cb86c754f884"
                  },
                  "quantity": {
                    "type": "number",
                    "example": 1
                  }
                }
              },
              "example": {
                "product_id": "f1498e34-b5fc-42bf-8fa0-cb86c754f884",
                "quantity": 1
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Stock reduced successfully"
                    },
                    "updatedBatches": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "batchId": {
                            "type": "string",
                            "example": "KAT25_031742559499"
                          },
                          "remainingStock": {
                            "type": "integer",
                            "example": 9
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          }
        }
      }
    },
...
  }
}